switch to go modules
