--- Gopkg.toml
@@ -1,36 +1,0 @@
-# Refer to https://github.com/golang/dep/blob/master/docs/Gopkg.toml.md
-# for detailed Gopkg.toml documentation.
-
-ignored = ["github.com/cayleygraph/cayley/internal/dock"]
-
-[[constraint]]
-  branch = "master"
-  name = "github.com/badgerodon/peg"
-
-[[constraint]]
-  branch = "master"
-  name = "github.com/dennwc/graphql"
-
-[[constraint]]
-  name = "github.com/go-kivik/pouchdb"
-  version = "1.3.5"
-
-[[constraint]]
-  branch = "master"
-  name = "github.com/golang/glog"
-
-[[constraint]]
-  branch = "master"
-  name = "github.com/linkeddata/gojsonld"
-
-[[constraint]]
-  branch = "v2"
-  name = "gopkg.in/mgo.v2"
-
-[[constraint]]
-  name = "gopkg.in/olivere/elastic.v5"
-  version = "5.0.58"
-
-[prune]
-  go-tests = true
-  unused-packages = true
