+++ docs/Contributing.md
@@ -5,21 +5,12 @@ Join our community on [discourse.cayley.io](https://discourse.cayley.io) or othe
 # Simply building Cayley
 
 ```
-# setup GOPATH
-mkdir -p ~/cayley && cd ~/cayley
-export GOPATH=`pwd`
-export PATH=$PATH:~/cayley/bin
-mkdir -p bin pkg
-
 # clone project
-mkdir -p src/github.com/cayleygraph
-cd src/github.com/cayleygraph
 git clone https://github.com/cayleygraph/cayley
 cd cayley
 
 # download dependencies
-curl https://raw.githubusercontent.com/golang/dep/master/install.sh | sh
-dep ensure --vendor-only
+go mod donwload
 
 # build the binary
 go build ./cmd/cayley
@@ -38,11 +29,13 @@ To run the web frontend, replace the "repl" command with "http"
 
 # Hacking on Cayley
 
-First, you'll need Go [(version 1.9.x or greater)](https://golang.org/doc/install) and a Go workspace. This is outlined by the Go team at http://golang.org/doc/code.html and is sort of the official way of going about it.
+First, you'll need Go [(version 1.11.x or greater)](https://golang.org/doc/install) and a Go workspace.
+This is outlined by the Go team at http://golang.org/doc/code.html and is sort of the official way of going about it.
 
 If you just want to build Cayley and check out the source, or use it as a library, a simple `go get github.com/cayleygraph/cayley` will work!
 
-But suppose you want to contribute back on your own fork (and pull requests are welcome!). A good way to do this is to set up your $GOPATH and then...
+But suppose you want to contribute back on your own fork (and pull requests are welcome!).
+A good way to do this is to set up your $GOPATH and then...
 
 ```
 mkdir -p $GOPATH/src/github.com/cayleygraph
@@ -68,8 +61,6 @@ With that in place, that folder will reflect your local fork, be able to take ch
 For iterating, it can be helpful to, from the directory, run
 
 ```
-curl https://raw.githubusercontent.com/golang/dep/master/install.sh | sh
-dep ensure --vendor-only
 go build ./cmd/cayley && ./cayley <subcommand> <your options>
 ```
 
@@ -79,9 +70,7 @@ Which will also resolve the relevant static content paths for serving HTTP.
 
 # Running Unit Tests
 
-First, `cd` into the `cayley` project folder.
-
-For Go 1.9 and onwards:
+First, `cd` into the `cayley` project folder and run:
 ```
 go test ./...
 ```
