--- imports_19.go
@@ -1,16 +1,0 @@
-// +build go1.9
-
-package cayley
-
-import (
-	"github.com/cayleygraph/cayley/graph"
-	"github.com/cayleygraph/cayley/graph/path"
-)
-
-type Iterator = graph.Iterator
-type QuadStore = graph.QuadStore
-type QuadWriter = graph.QuadWriter
-
-type Path = path.Path
-
-type Handle = graph.Handle
